Add logic to set EditText fields and to save changes to existing game type
Using overloading with makeIntent(), if the EditGameActivity boolean is true, it will grab the gametype's string name and grab it from gamemanager. From there, the EditTextFields will be set accordingly to edit and when the user saves, it will invoke a method to override fields in the existing GameType object.