Refactor camera and file permissions
The app on first startup will ask the user if they would can grant permissions. If the user declines, they must grant permissions within the options. If the user denies that again, it will not ask any more and the user will not be able to take custom photos. 2 ways to grant permissions: reinstall app or manually grant within the system settings for the app.